+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * :hammer: Network getIpAddress rename to getLocalIpAddress
 * :sparkles: Network add getLocalIpV4Address
 * :arrow_up: minSdkVersion rose to 16
+* :sparkles: Add StorageManagerCompat and StorageVolumeCompat
 
 ## v0.2.2
 * Upgrade Javax 0.6.4